Add files via upload
Movie Rating Prediction with Python

This project involves building a model to predict the ratings of movies based on various features such as genre, director, actors, and more. The objective is to use regression techniques to analyze historical movie data and accurately estimate the ratings given by users or critics.